fix(Image): demo拆解与规范
