+++ src/packages/image/doc.taro.md
@@ -18,19 +18,7 @@ import { Image } from '@nutui/nutui-react-taro';
 
 :::demo
 
-```tsx
-import React from "react";
-import { Image } from '@nutui/nutui-react-taro';
-
-const App = () => {
-  const src =
-    'https://storage.360buyimg.com/imgtools/e067cd5b69-07c864c0-dd02-11ed-8b2c-d7f58b17086a.png'
-  return <>
-    <Image src={src} height={200} />
-  </>
-}
-export default App;
-```
+<CodeBlock src='taro/demo1.tsx'></CodeBlock>
 
 :::
 
@@ -40,51 +28,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React from "react";
-import { Image, Row, Col } from '@nutui/nutui-react-taro';
-
-const App = () => {
-  const src =
-    'https://storage.360buyimg.com/imgtools/e067cd5b69-07c864c0-dd02-11ed-8b2c-d7f58b17086a.png'
-  return <>
-    <Row gutter={10}>
-      <Col span="8">
-        <Image
-          src={src}
-          mode="aspectFit"
-          width="80"
-          height="80"
-          radius="50%"
-        />
-      </Col>
-    </Row>
-    <Row gutter={10}>
-      <Col span="8">
-        <Image
-          src={src}
-          mode="scaleToFill"
-          width="80"
-          height="80"
-          radius="50%"
-        />
-      </Col>
-    </Row>
-    <Row gutter={10}>
-      <Col span="8">
-        <Image
-          src={src}
-          mode="scaleToFill"
-          width="80"
-          height="80"
-          radius="10px"
-        />
-      </Col>
-    </Row>
-  </>
-}
-export default App;
-```
+<CodeBlock src='taro/demo2.tsx'></CodeBlock>
 
 :::
 
@@ -94,34 +38,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React from "react";
-import { Image, Row, Col } from '@nutui/nutui-react-taro';
-import { Loading } from '@nutui/icons-react-taro';
-import {  View } from '@tarojs/components'
-
-const App = () => {
-  const src =
-    'https://storage.360buyimg.com/imgtools/e067cd5b69-07c864c0-dd02-11ed-8b2c-d7f58b17086a.png'
-  return <>
-    <Row gutter={10}>
-      <Col span="8">
-        <Image width="80" height="80" />
-        <View style={{ textAlign: 'left' }}>默认</View>
-      </Col>
-      <Col span="8">
-        <Image
-          width="80"
-          height="80"
-          loading={<Loading className="nut-icon-loading" />}
-        />
-        <View style={{ textAlign: 'left' }}>自定义</View>
-      </Col>
-    </Row>
-  </>
-}
-export default App;
-```
+<CodeBlock src='taro/demo3.tsx'></CodeBlock>
 
 :::
 
@@ -131,33 +48,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React from "react";
-import { Image, Row, Col } from '@nutui/nutui-react-taro';
-import { Failure } from '@nutui/icons-react-taro';
-import {  View } from '@tarojs/components'
-
-const App = () => {
-  return <>
-    <Row gutter={10}>
-      <Col span="8">
-        <Image src="https://x" width="80" height="80" />
-        <View style={{ textAlign: 'left' }}>默认</View>
-      </Col>
-      <Col span="8">
-        <Image
-          src="https://x"
-          width="80"
-          height="80"
-          error={<Failure />}
-        />
-        <View style={{ textAlign: 'left' }}>自定义</View>
-      </Col>
-    </Row>
-  </>
-}
-export default App;
-```
+<CodeBlock src='taro/demo4.tsx'></CodeBlock>
 
 :::
 
@@ -167,32 +58,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React from "react";
-import { Image } from '@nutui/nutui-react-taro';
-
-const App = () => {
-  return <>
-    <Image
-      src="http://m.360buyimg.com/babel/s181x181_jfs/t1/210178/19/10205/31538/619bbcd9E5071aed5/8e1b7eb632aeed49.png"
-      width="30"
-      height="30"
-      style={{ marginRight: '10px' }}
-      onError={() => {
-        console.log('image error')
-      }}
-    />
-    <div
-      style={{
-        width: '220px',
-      }}
-    >
-      雪纺衫女2021年春季新款洋气轻熟上衣
-    </div>
-  </>
-}
-export default App;
-```
+<CodeBlock src='taro/demo5.tsx'></CodeBlock>
 
 :::
 
@@ -200,32 +66,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React from "react";
-import { Cell, Col, Image } from '@nutui/nutui-react-taro';
-import { Failure } from '@nutui/icons-react-taro';
-
-const App = () => {
-  return <>
-    <Cell style={{ flexWrap: 'wrap' }}>
-      {[
-        'scaleToFill',
-        'aspectFit',
-        'aspectFill',
-        'widthFix',
-        'heightFix',
-      ].map((mode) => {
-        return (
-          <Col span="8" key={mode}>
-            <Image src={src} mode={mode as any} width="80" height="80" />
-          </Col>
-        )
-      })}
-    </Cell>
-  </>
-}
-export default App;
-```
+<CodeBlock src='taro/demo6.tsx'></CodeBlock>
 
 :::
 
@@ -233,36 +74,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React from "react";
-import { Cell, Col, Image } from '@nutui/nutui-react-taro';
-import { Failure } from '@nutui/icons-react-taro';
-
-const App = () => {
-  return <>
-    <Cell style={{ flexWrap: 'wrap' }}>
-      {[
-        'top',
-        'bottom',
-        'center',
-        'left',
-        'right',
-        'top left',
-        'top right',
-        'bottom left',
-        'bottom right',
-      ].map((mode) => {
-        return (
-          <Col span="8" key={mode}>
-            <Image src={src} mode={mode as any} width="80" height="80" />
-          </Col>
-        )
-      })}
-    </Cell>
-  </>
-}
-export default App;
-```
+<CodeBlock src='taro/demo7.tsx'></CodeBlock>
 
 :::
 
@@ -270,26 +82,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React from 'react'
-import { Image } from '@nutui/nutui-react-taro'
-import { ScrollView } from '@tarojs/components'
-
-const App = () => {
-  return <>
-    <ScrollView style={{ height: '350px' }} scrollY>
-      {[
-        ...new Array(3).fill(src),
-        'https://storage.360buyimg.com/jdc-article/welcomenutui.jpg',
-        'https://storage.360buyimg.com/jdc-article/NutUItaro2.jpg',
-      ].map((_) => {
-        return <Image key={_} src={_} lazyLoad width='100%' height='150' />
-      })}
-    </ScrollView>
-  </>
-}
-export default App
-```
+<CodeBlock src='taro/demo8.tsx'></CodeBlock>
 
 :::
